converted Add Activity to a Fragment
