+++ code/app/src/main/res/drawable/rounded_background.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
 <corners android:radius="27dp"/> <!-- Adjust radius for desired roundness -->
-<solid android:color="#78766a"/> <!-- Grey color matching your design -->
+<solid android:color="#78766a"/>
 </shape>
